PLUGINS/UCX: Move sglXfer into its own ucx_sgl.{h,cpp}
