+++ src/plugins/ucx/meson.build
@@ -22,6 +22,7 @@ ucx_backend_sources = ['config.cpp',
                        'mem_list.cpp',
                        'rkey.cpp',
                        'ucx_backend.cpp',
+                       'ucx_sgl.cpp',
                        'ucx_enums.cpp',
                        'ucx_plugin.cpp',
                        'ucx_utils.cpp']
